Store new read_percent when hitting the back button from an article
Still has a problem because the read list stores the read_percent as
data on the row.  To re-open the article with the new read_percent
value, you have to go back one more time to the main menu.  Ideally,
we'd pull the article data out of the data store instead of out of the
row itself, that way we'll always have the most recent data.  I'm not
really up to a change like this myself.